Task_3/requirements.txt
black==24.10.0
Django==5.1.5
django-constance==4.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Jinja2==3.1.4
jose==1.0.0
Pygments==2.18.0
PyJWT==2.10.1
pylint==3.3.2
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
redis==5.2.1


